pyproject.toml
[build-system]
requires = ["scikit-build-core", "pybind11"]
build-backend = "scikit_build_core.build"

[project]
name = "lwl"
dynamic = ["version"]
readme = "README.md"

dependencies = ["numpy==1.24.4", 
                "pyyaml==6.0",
                "open3d==0.13.0",
                "seaborn==0.13.2",
                "torch==1.12.1+cu113",
                "torcheval==0.0.7",
                "wandb", "plotly"]

[tool.scikit-build]
metadata.version.provider = "scikit_build_core.metadata.setuptools_scm"
sdist.include = ["lwl/_version.py"]
cmake.build-type = "Release"

[tool.setuptools_scm]
write_to = "lwl/_version.py"